(#4) Move project in a separate folder
