+++ Alice.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29025.244
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Alice.Discord", "Alice.Discord\Alice.Discord.csproj", "{3E5BABF0-143D-4A06-8D79-73441114ED11}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3E5BABF0-143D-4A06-8D79-73441114ED11}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3E5BABF0-143D-4A06-8D79-73441114ED11}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3E5BABF0-143D-4A06-8D79-73441114ED11}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3E5BABF0-143D-4A06-8D79-73441114ED11}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A8BA3110-C0BF-4E0D-9ED9-77FCEB0207EF}
+	EndGlobalSection
+EndGlobal
